rename cosmos query relayer to neutron query relayer
