+++ .gitignore
@@ -24,8 +24,8 @@ go.work
 
 keys
 
-cmd/cosmos_query_relayer/test.go
-/cosmos_query_relayer
+cmd/neutron_query_relayer/test.go
+/neutron_query_relayer
 
 .env
