+++ Makefile
@@ -1,11 +1,11 @@
 dev:
-	go run ./cmd/cosmos_query_relayer/
+	go run ./cmd/neutron_query_relayer/
 
 test:
 	 go test ./...
 
 build:
-	go build -a -o cosmos_query_relayer ./cmd/cosmos_query_relayer/*.go
+	go build -a -o neutron_query_relayer ./cmd/neutron_query_relayer/*.go
 
 build-docker:
-	docker build . -t neutron-org/cosmos-query-relayer --ssh default
+	docker build . -t neutron-org/neutron-query-relayer --ssh default
