+++ internal/proof/proof_impl/proof_impl.go
@@ -1,8 +1,8 @@
 package proof_impl
 
 import (
-	"github.com/neutron-org/cosmos-query-relayer/internal/proof"
-	"github.com/neutron-org/cosmos-query-relayer/internal/relay"
+	"github.com/neutron-org/neutron-query-relayer/internal/proof"
+	"github.com/neutron-org/neutron-query-relayer/internal/relay"
 )
 
 type ProoferImpl struct {
